fix: move validation indication to subheader to be app-development specific
